+++ src/index.ts
@@ -4,7 +4,7 @@ export * from "./amigaHunkParser";
 export * from "./disassembly/cpuDisassembler";
 export * from "./disassembly/copperDisassembler";
 export * from "./disassembly/disassembledFile";
-export * from "./customRegisters";
+export * from "./hardware";
 export { BreakpointStorage } from "./breakpoints";
 export { MemoryFormat } from "./program";
 export { NumberFormat } from "./utils/strings";